requirements.txt
pandas 
numpy
tensorflow-cpu
scikit-learn
scipy
webbrowser
pillow
streamlit
streamlit_oauth
python-dotenv
matplotlib
seaborn